refactor:#265 client meal 리팩토링
